requirements.txt
pydantic
PyMuPDF
openai
arxiv
python-dotenv
chromadb
sentence-transformers
tiktoken
tqdm